+++ tests/test_keras_integration.py
@@ -14,6 +14,7 @@ from huggingface_hub.keras_mixin import (
     push_to_hub_keras,
     save_pretrained_keras,
 )
+from huggingface_hub.utils import logging
 
 from .testing_constants import ENDPOINT_STAGING, PASS, USER
 from .testing_utils import set_write_permission_and_retry
@@ -22,6 +23,8 @@ from .testing_utils import set_write_permission_and_retry
 def repo_name(id=uuid.uuid4().hex[:6]):
     return "keras-repo-{0}-{1}".format(id, int(time.time() * 10e3))
 
+
+logger = logging.get_logger(__name__)
 
 WORKING_REPO_SUBDIR = f"fixtures/working_repo_{__name__.split('.')[-1]}"
 WORKING_REPO_DIR = os.path.join(
@@ -65,10 +68,11 @@ else:
 @require_tf
 class HubMixingTestKeras(unittest.TestCase):
     def tearDown(self) -> None:
-        try:
+        if os.path.exists(WORKING_REPO_DIR):
             shutil.rmtree(WORKING_REPO_DIR, onerror=set_write_permission_and_retry)
-        except FileNotFoundError:
-            pass
+        logger.info(
+            f"Does {WORKING_REPO_DIR} exist: {os.path.exists(WORKING_REPO_DIR)}"
+        )
 
     @classmethod
     def setUpClass(cls):
